requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
tortoise-orm[asyncpg]>=v0.25.0,<v0.26.0
aerich[toml]==v0.8.2
strawberry-graphql[fastapi]>=0.266.1, <0.267.0
